Split up tests for better readability
