+++ test.py
@@ -1,0 +1,5 @@
+import unittest
+
+if __name__ == '__main__':
+    testsuite = unittest.TestLoader().discover('.')
+    unittest.TextTestRunner(verbosity=1).run(testsuite)
